Option for loading .cl sources from a given path
Add option for loading .cl sources from a different path on RenderGirl
core. Interfaces will have to provided where the sources are located,
since each interface may have a different installation procedure and
files organization.

It would be interesting to have a list of path (can it be colon
separated, such as unix paths), but I don't want to overdesign the core
for a feature I don't need it now.