Blender transmits scene data to the Core
BlenderPlugin will now transmits geometry information to the Core using
the expected format from RenderGirl. To fetch geometric information, we
use the experimental bmesh API from Blender. The bmesh has a couple of
intersting mapping, such as faces -> vert, so we used this to recreated
the mapping present on the Core (a faces buffer pointing to vertex
buffer, only one per scene group). Each Blender object is translated to
a scene group. A "triangulate" modifier was used to triangulate the mesh
prior transmission.