Fix camera computations on RenderGirlBlender
The camera system on RenderGirl currently does not provide a good API
for providing camera attributes. Until this commit, a "look at" vector
was necessary to perform any rendering. Since Blender does not provide
one, I had to add an option for the Core to use a provided direction
from the camera.

A major rewrite on the camera API will be needed in the future, with a
specialized class responsible for receiving camera information (with or
without look at) and taking care of setting up the camera within the
scene.

Known issue: 3D viewport is not generated properly when aspect ratio is
different than 1:1.