SceneGroup perform local to global transformation
Moves local to global vertex transformation from SceneManager to
SceneGroup class. This introduces a new object m_local_vertices
parameter which states if the vertex information within the SceneGroup
is in local space or global space. For now, the class will denies any
new vertice if the object is in global mode.