Add BVH creation on CPU side
Implemented the "Create" function from the BVH class which recursivly
creates the BVH structure from a given scene. For now, the data is
created by it's not used.

Change the container type storing the groups on SceneManager from a list
to a vector, to better iterate over the objects when building the BVHs.